initial handling of examples. Work in progess

+++ fabfed/provider/chi/chi_slice.py
@@ -1,31 +1,6 @@
-#!/usr/bin/env python3
-# MIT License
-#
-# Copyright (c) 2020 RENCI NRIG
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-#
-# Author Komal Thareja (kthare10@renci.org)
 import logging
 
 from fabfed.model import Slice
-from fabfed.model.state import SliceState
 from fabfed.provider.chi.chi_network import ChiNetwork
 from fabfed.provider.chi.chi_node import ChiNode
 from fabfed.util.constants import Constants
@@ -48,12 +23,13 @@ class ChiSlice(Slice):
         pool_start = resource.get(Constants.RES_NET_POOL_START, None)
         pool_end = resource.get(Constants.RES_NET_POOL_END, None)
         gateway = resource.get(Constants.RES_NET_GATEWAY, None)
-        stitch_provider = resource.get(Constants.RES_NET_STITCH_PROV, None)
+        stitch_providers = resource.get(Constants.RES_NET_STITCH_PROVS, list())
+        assert stitch_providers
 
         net_name = f'{self.name}-{resource.get(Constants.RES_NAME_PREFIX)}'
         net = ChiNetwork(label=label, name=net_name, site=site, logger=self.logger,
                          slice_name=self.name, subnet=subnet, pool_start=pool_start, pool_end=pool_end,
-                         gateway=gateway, stitch_provider=stitch_provider,
+                         gateway=gateway, stitch_provider=stitch_providers[0],
                          project_name=self.project_name)
         self._networks.append(net)
 
@@ -123,26 +99,31 @@ class ChiSlice(Slice):
         self.slice_created = True
 
     # noinspection PyTypeChecker
-    def destroy(self, *, slice_state: SliceState):
-        node_states = slice_state.node_states
+    def delete_resource(self, *, resource: dict):
+        # name = resource['name']
+        site = resource[Constants.RES_SITE]
+        label = resource.get(Constants.LABEL)
+        rtype = resource.get(Constants.RES_TYPE)
 
-        for node_state in node_states:
-            name = node_state.attributes['name']
-            site = node_state.attributes[Constants.RES_SITE]
-            self.logger.debug(f"Deleting node: {name} at site {site}")
-            node = ChiNode(label=node_state.label, name=name, image=None, site=site, flavor=None, logger=self.logger,
-                           key_pair=None, slice_name=self.name, network=None,
-                           project_name=self.project_name)
-            node.delete()
-
-        network_states = slice_state.network_states
-
-        for network_state in network_states:
-            name = network_state.attributes['name']
-            site = network_state.attributes[Constants.RES_SITE]
-            self.logger.debug(f"Deleting network: {name} at site {site}")
-            net = ChiNetwork(label=network_state.label, name=name, site=site, logger=self.logger,
+        if rtype == Constants.RES_TYPE_NETWORK.lower():
+            net_name = f'{self.name}-{resource.get(Constants.RES_NAME_PREFIX)}'
+            self.logger.debug(f"Deleting network: {net_name} at site {site}")
+            net = ChiNetwork(label=label, name=net_name, site=site, logger=self.logger,
                              slice_name=self.name, subnet=None, pool_start=None, pool_end=None,
                              gateway=None, stitch_provider=None,
                              project_name=self.project_name)
             net.delete()
+            self.logger.info(f"Deleted network: {net_name} at site {site}")
+        else:
+            node_count = resource.get(Constants.RES_COUNT, 1)
+
+            for n in range(0, node_count):
+                node_name_prefix = resource.get(Constants.RES_NAME_PREFIX)
+                node_name = f"{node_name_prefix}{n}"
+                self.logger.debug(f"Deleting node: {node_name} at site {site}")
+
+                node = ChiNode(label=label, name=node_name, image=None, site=site, flavor=None, logger=self.logger,
+                               key_pair=None, slice_name=self.name, network=None,
+                               project_name=self.project_name)
+                node.delete()
+                self.logger.info(f"Deleted node: {node_name} at site {site}")
